requirements.txt
pillow
tk
pytesseract
requests
google-generativeai
python-dotenv
opencv-python